@@ -488,8 +494,8 @@
         <artifactId>maven-compiler-plugin</artifactId>
         <inherited>true</inherited>
         <configuration>
-          <source>${main.java.version}</source>
-          <target>${main.java.version}</target>
+          <source>21</source>
+          <target>21</target>
         </configuration>
         <executions>
           <!-- Ensure main source tree compiles to Java ${main.java.version} bytecode. -->
@@ -500,8 +506,8 @@
             </goals>
             <phase>compile</phase>
             <configuration>
-              <source>${main.java.version}</source>
-              <target>${main.java.version}</target>
+              <source>21</source>
+              <target>21</target>
             </configuration>
           </execution>
           <execution>
@@ -511,8 +517,8 @@
             </goals>
             <phase>test-compile</phase>
             <configuration>
-              <source>${latest.java.version}</source>
-              <target>${latest.java.version}</target>
+              <source>21</source>
+              <target>21</target>
             </configuration>
           </execution>
         </executions>
@@ -779,7 +777,7 @@
             <groupId>org.apache.maven.plugins</groupId>
             <artifactId>maven-javadoc-plugin</artifactId>
             <configuration>
-              <source>8</source>
+              <source>21</source>
               <excludePackageNames>com.querydsl.examples.kotlin.*</excludePackageNames>
               <outputDirectory>${project.build.directory}/dist/apidocs</outputDirectory>
               <reportOutputDirectory>${project.build.directory}/dist/apidocs</reportOutputDirectory>
