@@ -16,7 +16,7 @@ package com.querydsl.core.types.dsl;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.querydsl.core.types.ConstantImpl;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 public class ArrayPathTest {
 
