@@ -4,15 +4,15 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Arrays;
 import java.util.HashSet;
-import org.junit.Before;
 import org.junit.Ignore;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 public class BeanMapTest {
 
   private BeanMap beanMap;
 
-  @Before
+  @BeforeEach
   public void setUp() {
     beanMap = new BeanMap(new Entity());
   }
