@@ -18,14 +18,7 @@ package com.querydsl.mongodb.document;
 import com.mongodb.DBRef;
 import com.querydsl.core.types.*;
 import com.querydsl.mongodb.MongodbOps;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.LinkedHashSet;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Map;
-import java.util.Queue;
-import java.util.Set;
+import java.util.*;
 import java.util.regex.Pattern;
 import org.bson.BsonJavaScript;
 import org.bson.BsonRegularExpression;
